added the comment.py file to add some routes for the comment model added test files for time, showtime, booking and comment made some changes to the naming conventions of the routes and the functions in the showTime and time py files added the admin functionality to the time and show time delete routes

+++ app/routers/showTime.py
@@ -8,14 +8,12 @@ router = APIRouter(prefix="/showTime", tags=["ShowTimes"])
 
 
 @router.post("/")
-def create_time(
+def create_showtime(
     time: schemas.ShowTimes,
     db: Session = Depends(get_db),
     current_user: int = Depends(oauth2.get_current_user),
 ):
-    admin_status = current_user.admin
-    print(admin_status)
-    if admin_status == "true":
+    if current_user.admin == "true":
         new_time = models.ShowTime(**time.dict())
         db.add(new_time)
         db.commit()
@@ -29,21 +27,31 @@ def create_time(
 
 
 @router.get("/")
-def get_times(db: Session = Depends(get_db)):
+def get_showtimes(db: Session = Depends(get_db)):
     movies = db.query(models.ShowTime).all()
     return movies
 
 
 @router.delete("/{id}")
-def delete_showTime(id: int, db: Session = Depends(get_db)):
+def delete_showTime(
+    id: int,
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+):
     time_query = db.query(models.ShowTime).filter(models.ShowTime.id == id)
 
-    if time_query.first() == None:
+    if current_user.admin == "true":
+        if time_query.first() == None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"post with id: {id} does not exist",
+            )
+        time_query.delete(synchronize_session=False)
+        db.commit()
+    else:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"post with id: {id} does not exist",
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=f"You are not allowed to perform this action because you are not an admin",
         )
-    time_query.delete(synchronize_session=False)
-    db.commit()
 
     return Response(status_code=status.HTTP_204_NO_CONTENT)
